fix(jscad-worker): fixed other place with bad CSG/CAG type detection
