+++ src/core/jscad-worker.js
@@ -1,19 +1,17 @@
 // jscad-worker.js
 //
 // == OpenJSCAD.org, Copyright (c) 2013-2016, Licensed under MIT License
-
-const { CAG, CSG } = require('@jscad/csg')
 const oscad = require('@jscad/scad-api')
 
 const createJscadFunction = require('./jscad-function')
 const { toArray } = require('../utils/misc')
+const {isCSG, isCAG} = require('./utils')
 
 /**
  * Create an worker (thread) for processing the JSCAD script into CSG/CAG objects
  */
 module.exports = function (self) {
   self.onmessage = function (e) {
-    var r = {cmd: 'error', txt: 'try again'}
     if (e.data instanceof Object) {
       var data = e.data
       if (data.cmd === 'render') {
@@ -25,7 +23,7 @@ module.exports = function (self) {
         let objects = func(parameters, include, globals)
         objects = toArray(objects)
           .map(function (object) {
-            if (object instanceof CAG || object instanceof CSG) {
+            if (isCSG(object) || isCAG(object)) {
               return object.toCompactBinary()
             }
           })
